Attend to review comments
* Use usize instead of u64 in PingBodyOverflow error;
* Be explicit about bytes unit in ping body size comments;
* Consume pending_pings array in correct order;
* Use meaningful value for body_max_size request tests;

+++ glean-core/src/error.rs
@@ -62,7 +62,7 @@ pub enum ErrorKind {
     NotInitialized,
 
     /// Ping request body size overflowed
-    PingBodyOverflow(u64),
+    PingBodyOverflow(usize),
 }
 
 /// A specialized [`Error`] type for this crate's operations.
